trying to rewrite the tip message implementation! It threw error when trying to run old code.
reading documentation for teloxide, and trying to find a way to obtain recipient ID for bot to send a message!

An idea I wrote down in commented code, lines 38-56, outlines my idea for trying to obtain recipient ID -> user must be in a channel that the bot has admin rights to!